fix(memory): move getLogsDbPath to its own module to dodge a Bun link bug
CI failed with "Export named 'getLogsDbPath' not found in module
platform.ts" in test files with large import graphs (heartbeat,
background-wake). The export is valid (it resolves fine when platform.ts
is imported in isolation), but adding a new export to platform.ts — a
module imported very early and widely — and consuming it from a
low-level module (db-connection.ts) across the daemon's large, cyclic
import graph trips a Bun module-linker ordering bug.

Fix: keep platform.ts byte-identical to before and put getLogsDbPath in
its own leaf module (util/logs-db-path.ts). db-connection, db status, and
the test import from there. Also drop the static logger import added to
db-connection.ts in favor of a lazy dynamic import on the error path, so
this deliberately import-light module doesn't pull in pino.
